[Crom] Chapter 6. QueryDSL 추가학습

+++ src/main/java/ssu/cromi/umc9th/domain/user/dto/UserProfileDto.java
@@ -10,5 +10,5 @@ public class UserProfileDto {
     private String nickname;
     private String email;
     private String phone;
-    private Integer userPoint;
+    private Long userPoint;
 }
